fix(auth): deduplicate role hierarchy and JWT key building, fix minor backend issues
- RoleHierarchy helper (package-private) extrait le switch de hiérarchie
  partagé entre RoleCreationPolicy et RoleDeactivationPolicy
- JwtKeyFactory helper (package-private) extrait la validation+build de
  SecretKey commune à JwtService et JwtValidationService
- CookieService : multiplication expiryMinutes * 60 en long pour cohérence
  avec JwtService
- RegisterHandler : email normalisé en minuscules (Locale.ROOT) à la
  création — empêche Alice@test.com et alice@test.com de coexister
- Liquibase : changeset 005 supprime idx_refresh_tokens_token_hash,
  redondant avec la contrainte unique existante
- ArchRulesTest : commentaire explicatif sur la règle web_controllers triviale
- AuthControllerIT : test 429 ajouté sur POST /auth/refresh

+++ src/main/java/com/synapsys/api/auth/infrastructure/security/CookieService.java
@@ -21,7 +21,7 @@ public class CookieService {
 
     public CookieService(SynapsysProperties properties) {
         this.secure = properties.cookie().secure();
-        this.accessMaxAgeSeconds = properties.jwt().expiryMinutes() * 60;
+        this.accessMaxAgeSeconds = (int) (properties.jwt().expiryMinutes() * 60L);
         this.refreshMaxAgeSeconds = properties.refreshToken().expiryDays() * 86_400;
     }
